requirements.txt
wheel==0.34.2
Django==2.2.19
django-currentuser==0.4.2
django-constance==2.4.0
django-environ==0.4.5
django-localflavor==2.2
django-localflavor-br==1.0
django-timezone-field==3.0
django-wkhtmltopdf==3.2.0
djangorestframework==3.10.2
goslate==1.5.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
django-reversion==3.0.4
django-automated-logging==6.1.0
django-log-viewer==1.1.2
git+git://github.com/marcelomdevasconcellos/django-adminlte-ui.git#egg=django-adminlte-ui #original: django-adminlte-ui==1.5.1
django-filter==2.2.0
django-picklefield
django-jsonfield==1.4.0
unidecode==1.1.1
django_extensions==3.0.3
pydotplus==2.0.2
graphviz==0.14.1
pandas==1.2.0
xlrd==2.0.1
openpyxl==3.0.6
termcolor==1.1.0
django-ajax-selects==2.1.0 #https://dev.to/thearjun/implement-autocomplete-in-django-3h20
django-admin-rangefilter==0.7.0 # https://github.com/silentsokolov/django-admin-rangefilter
pyexcel-ods3==0.6.0
requests==2.25.1
